Move util function from templatetags
